# ui/src/lib/generated/openapi.json
{
  "components": {
    "schemas": {
...
      "ClassList": {
        "additionalProperties": false,
        "properties": {
          "classes": {
            "items": {
              "type": "string"
            },
            "title": "Classes",
            "type": "array"
          },
          "kind": {
            "title": "Kind",
            "type": "string"
          }
        },
        "required": [
          "kind",
          "classes"
        ],
        "title": "ClassList",
        "type": "object"
      },
...
      "DownloadRequest": {
        "properties": {
          "repo_id": {
            "description": "Hub repo to download, e.g. org/model",
            "title": "Repo Id",
            "type": "string"
          }
        },
        "required": [
          "repo_id"
        ],
        "title": "DownloadRequest",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "ModelDeleted": {
        "additionalProperties": false,
        "properties": {
          "deleted": {
            "title": "Deleted",
            "type": "boolean"
          },
          "freed": {
            "description": "Bytes freed.",
            "title": "Freed",
            "type": "integer"
          },
          "repo_id": {
            "title": "Repo Id",
            "type": "string"
          }
        },
        "required": [
          "repo_id",
          "deleted",
          "freed"
        ],
        "title": "ModelDeleted",
        "type": "object"
      },
...
      "PipelineDescription": {
        "additionalProperties": false,
        "properties": {
          "accepts_kwargs": {
            "title": "Accepts Kwargs",
            "type": "boolean"
          },
          "compatibles": {
            "description": "Scheduler classes this one can swap with.",
            "items": {
              "type": "string"
            },
            "title": "Compatibles",
            "type": "array"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "parameters": {
            "items": {
              "$ref": "#/components/schemas/PipelineParameter"
            },
            "title": "Parameters",
            "type": "array"
          },
          "summary": {
            "title": "Summary",
            "type": "string"
          }
        },
        "required": [
          "name",
          "summary",
          "accepts_kwargs",
          "parameters"
        ],
        "title": "PipelineDescription",
        "type": "object"
      },
...
    }
  },
  "info": {
    "description": "Declarative diffusers workflows over HTTP: queue a job, stream its progress, fetch what it saved.",
    "title": "diffusers-workflow",
    "version": "0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/classes": {
      "get": {
        "description": "Class names of one kind (pipelines, models, schedulers,\nquantization) - the pickers' data source.",
        "operationId": "classes_api_classes_get",
        "parameters": [
          {
            "in": "query",
            "name": "kind",
            "required": true,
            "schema": {
              "title": "Kind",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ClassList"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Classes"
      }
    },
    "/api/classes/{name}": {
      "get": {
        "description": "A class's argument schema: target=call reads __call__, init reads\n__init__, load reads from_pretrained plus the curated loading knobs.",
        "operationId": "class_description_api_classes__name__get",
        "parameters": [
          {
            "in": "path",
            "name": "name",
            "required": true,
            "schema": {
              "title": "Name",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "target",
            "required": false,
            "schema": {
              "default": "init",
              "title": "Target",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PipelineDescription"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Class Description"
      }
    },
...
    "/api/guides/{name}": {
      "get": {
        "description": "One guide from /api/guides, whole or one section of it. A\nsection name is matched loosely - case and punctuation dropped -\nso a heading copied approximately still resolves, and also reaches\na `###` subsection not listed at the top level, by its own heading\nor by a term inside it. An unknown name or section is a 404 whose\ndetail lists what exists.",
        "operationId": "get_guide_api_guides__name__get",
        "parameters": [
          {
            "in": "path",
            "name": "name",
            "required": true,
            "schema": {
              "title": "Name",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "section",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Section"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Guide"
      }
    },
...
    "/api/models": {
      "delete": {
        "description": "Delete every cached revision of one repo from the hub cache.\n\nRefused while a job is running or queued: the worker may be reading\nexactly the files a delete would remove out from under it.",
        "operationId": "delete_cached_model_api_models_delete",
        "parameters": [
          {
            "in": "query",
            "name": "repo",
            "required": true,
            "schema": {
              "title": "Repo",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ModelDeleted"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Delete Cached Model"
      },
      "get": {
        "description": "What the Hugging Face hub cache holds, largest repo first.",
        "operationId": "get_models_api_models_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ModelCache"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "summary": "Get Models"
      }
    },
    "/api/models/download": {
      "post": {
        "description": "Start a background snapshot download into the hub cache.",
        "operationId": "start_download_api_models_download_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DownloadRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ModelDownload"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Start Download"
      }
    },
...
    "/api/models/downloads/{download_id}/cancel": {
      "post": {
        "description": "Request cancellation; takes effect at the next progress tick.\nPartial files stay in the cache and resume on a retry.",
        "operationId": "cancel_download_api_models_downloads__download_id__cancel_post",
        "parameters": [
          {
            "in": "path",
            "name": "download_id",
            "required": true,
            "schema": {
              "title": "Download Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ModelDownload"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Cancel Download"
      }
    },
...
    "/api/pipelines/{name}": {
      "get": {
        "description": "A pipeline's __call__ argument schema, for form generation.",
        "operationId": "pipeline_description_api_pipelines__name__get",
        "parameters": [
          {
            "in": "path",
            "name": "name",
            "required": true,
            "schema": {
              "title": "Name",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PipelineDescription"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Pipeline Description"
      }
    },
...
    "/api/schema": {
      "get": {
        "description": "The workflow JSON schema, for schema-aware JSON editing.\n\n`?section=` answers one part of it - `steps`, `pipelines`, `tasks`,\n`result`, `variables` or `configuration` - as\n`{section, sections, elsewhere, schema}`, for a reader that wants\nthe shape of a result block and not 36 KB of quantization configs\n(#101). Additive: the no-argument call is the whole schema, as it\nwas. An unknown section is a 404 naming the ones that exist.",
        "operationId": "workflow_schema_api_schema_get",
        "parameters": [
          {
            "in": "query",
            "name": "section",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Section"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Workflow Schema"
      }
    },
    "/api/server": {
      "get": {
        "description": "How this server is reachable, for the UI's Server page: what it\nis bound to, whether a token is needed, whether MCP is mounted, and\nthe addresses another machine could name it by.\n\nNo URL is composed here - the caller pairs an address with `port`\nand `mcp.path` - and the token itself is never reported in any\nform, only whether one is required. An interface enumeration\nfailure is not a server failure: `addresses` comes back empty.\n\n`directories` is scoped to the `?workspace=` a caller names (or the\nsession's own pin, via `_scoped`) - a mounted `download_output`\nconfines a write to *that* workspace's output tree, so reporting\nthe server's own default here regardless of the selector sent a\ncaller pinned elsewhere writing into `default` without any error (#389).",
        "operationId": "server_info_api_server_get",
        "parameters": [
          {
            "in": "query",
            "name": "workspace",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Workspace"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ServerInfo"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Server Info"
      }
    },
...
    "/api/system/diffusers/update": {
      "post": {
        "description": "Upgrade diffusers in the background: GitHub HEAD by default, a\npinned commit when `commit` is given, or a revert to the last\nknown-good published release when `revert` is true.\n\nRefused while a job is running or queued: pip replacing package\nfiles under a loaded pipeline is the model-delete hazard in another\nform. On success the idle worker is shut down so the next job\nimports the new version.",
        "operationId": "update_diffusers_api_system_diffusers_update_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateDiffusersRequest",
                "default": {
                  "revert": false
                }
              }
            }
          }
        },
        "responses": {
          "202": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DiffusersStatus"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Update Diffusers"
      }
    },
...
    "/api/tasks/{command}": {
      "get": {
        "description": "A task command's argument schema - the registered implementation\nfunction's real signature, in the same shape as a class description.",
        "operationId": "get_task_api_tasks__command__get",
        "parameters": [
          {
            "in": "path",
            "name": "command",
            "required": true,
            "schema": {
              "title": "Command",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PipelineDescription"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Task"
      }
    },
...
  }
}